+++ internal/provider/schema_util.go
@@ -66,6 +66,37 @@ func MustAddMountMigrationSchema(r *schema.Resource, customStateUpgrade bool) *s
 	return r
 }
 
+func MustAddSecretsSyncCommonSchema(r *schema.Resource) *schema.Resource {
+	MustAddSchema(r, map[string]*schema.Schema{
+		consts.FieldType: {
+			Type:        schema.TypeString,
+			Computed:    true,
+			Description: "Type of secrets destination.",
+			ForceNew:    true,
+		},
+		consts.FieldSecretNameTemplate: {
+			Type:        schema.TypeString,
+			Optional:    true,
+			Computed:    true,
+			Description: "Template describing how to generate external secret names.",
+		},
+	})
+
+	return r
+}
+
+func MustAddSecretsSyncCloudSchema(r *schema.Resource) *schema.Resource {
+	MustAddSchema(r, map[string]*schema.Schema{
+		consts.FieldCustomTags: {
+			Type:        schema.TypeMap,
+			Optional:    true,
+			Description: "Custom tags to set on the secret managed at the destination.",
+		},
+	})
+
+	return MustAddSecretsSyncCommonSchema(r)
+}
+
 func GetNamespaceSchema() map[string]*schema.Schema {
 	return map[string]*schema.Schema{
 		consts.FieldNamespace: {