+++ vault/provider.go
@@ -729,6 +729,34 @@ var (
 			Resource:      UpdateSchemaResource(samlAuthBackendRoleResource()),
 			PathInventory: []string{"/auth/saml/role/{name}"},
 		},
+		"vault_secrets_sync_config": {
+			Resource:      UpdateSchemaResource(secretsSyncConfigResource()),
+			PathInventory: []string{"/sys/sync/config"},
+		},
+		"vault_secrets_sync_aws_destination": {
+			Resource:      UpdateSchemaResource(awsSecretsSyncDestinationResource()),
+			PathInventory: []string{"/sys/sync/destinations/aws-sm/{name}"},
+		},
+		"vault_secrets_sync_azure_destination": {
+			Resource:      UpdateSchemaResource(azureSecretsSyncDestinationResource()),
+			PathInventory: []string{"/sys/sync/destinations/azure-kv/{name}"},
+		},
+		"vault_secrets_sync_gcp_destination": {
+			Resource:      UpdateSchemaResource(gcpSecretsSyncDestinationResource()),
+			PathInventory: []string{"/sys/sync/destinations/gcp-sm/{name}"},
+		},
+		"vault_secrets_sync_gh_destination": {
+			Resource:      UpdateSchemaResource(githubSecretsSyncDestinationResource()),
+			PathInventory: []string{"/sys/sync/destinations/gh/{name}"},
+		},
+		"vault_secrets_sync_vercel_destination": {
+			Resource:      UpdateSchemaResource(vercelSecretsSyncDestinationResource()),
+			PathInventory: []string{"/sys/sync/destinations/vercel-project/{name}"},
+		},
+		"vault_secrets_sync_association": {
+			Resource:      UpdateSchemaResource(secretsSyncAssociationResource()),
+			PathInventory: []string{"/sys/sync/destinations/{type}/{name}/associations/set"},
+		},
 	}
 )